backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# JWT token verification
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database (async PostgreSQL)
asyncpg==0.29.0

# Environment variables
python-dotenv==1.0.0

# HTTP client (for Supabase API calls if needed)
httpx==0.26.0

# CORS
fastapi-cors==0.0.6

# Pydantic for data validation (comes with FastAPI but explicit version)
pydantic==2.5.3
pydantic-settings==2.1.0

# Email sending
sendgrid==6.11.0

# PDF generation
reportlab==4.1.0

# Task scheduling
apscheduler==3.10.4

# Testing (dev dependencies)
pytest==7.4.4
pytest-asyncio==0.23.3
